partial refactor to remove some hackiness
